Migrate cattaraugus (rotated-SOVC x-anchor geometry) — completes sovc_table family
Reuses ny2024_rpp_parser rotated-diagonal helpers; emits reassembled anchor
names verbatim. All 31 migrated counties verify BYTE-IDENTICAL.
